Supply BUNDLE_FROZEN=true to prevent Gemfile.lock modifications.
In Ruby 3.3, bundler attempts to update Gemfile.lock metadata on bundle
install unless --frozen is used. Since the app is effectively read-only
in the current dockerfile as the owner is not changed, attempts to
update metadata of the file will be met with "permission denied" errors.

Supplying BUNDLE_FROZEN to ENV is the same as setting the --frozen on
CLI, and prevents bundler from updating Gemfile.lock at all and treat it
only read-only.

Additionally, to make sure we aren't using new CLI or approaches that
might be incompatible in future or past rubies, we supply an ENV variable
that gets read by bundler shipped with Ruby 3.3.

This can be met, using Dockerfile before this commit and running the
container build.
```
$ bundle install
<...snip...>
Installing sinatra-activerecord 2.0.22
There was an error while trying to write to `/opt/app-root/src/Gemfile.lock`. It
is likely that you need to grant write permissions for that path.
```